defer cythonization until cuda-bindings is installed
